+++ src/types/assets.ts
@@ -1,12 +1,5 @@
 import { AssetType } from "stellar-base";
 import { Horizon } from "./../horizon_api";
-
-export interface OfferAsset {
-  asset_type: AssetType;
-  asset_code?: string;
-  asset_issuer?: string;
-}
-
 export interface AssetRecord extends Horizon.BaseResponse {
   asset_type: AssetType.credit4 | AssetType.credit12;
   asset_code: string;